+++ analytic/analytic_for.py
@@ -1,0 +1,2 @@
+# TODO: new api
+from netharn.analytic.analytic_for import *