+++ analytic/output_shape_for.py
@@ -1,0 +1,2 @@
+# TODO: new api
+from netharn.analytic.output_shape_for import *