+++ netharn/models/dual_path_net.py
@@ -9,7 +9,7 @@ References:
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-# from netharn.output_shape_for import OutputShapeFor
+# from netharn.analytic.output_shape_for import OutputShapeFor
 from netharn.layers import ConvNorm2d
 
 # __all__ = ['DPN']